terrabot_design4_description/requirements.txt
absl-py==1.4.0
actionlib==1.14.0
ale-py==0.8.1
angles==1.9.13
anyio==3.7.0
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.1
attrs==19.3.0
autobahn==17.10.1
Automat==0.8.0
AutoROM==0.6.1
AutoROM.accept-rom-license==0.6.1
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.4
bondpy==1.8.6
breezy==3.0.2
Brlapi==0.7.0
cachetools==5.3.1
camera-calibration==1.17.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
catkin-pkg==0.5.2
catkin-pkg-modules==0.5.2
cbor==1.0.0
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
Click==7.0
cloudpickle==2.2.1
cmake==3.26.4
colorama==0.4.3
comm==0.1.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
controller-manager==0.19.6
controller-manager-msgs==0.19.6
coverage==4.5.2
cryptography==2.8
cupshelpers==1.0
cv-bridge==1.16.2
cycler==0.10.0
Cython==0.29.14
dbus-python==1.2.16
debugpy==1.6.7
decorator==5.1.1
defer==1.0.6
defusedxml==0.6.0
Deprecated==1.2.7
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
distro==1.4.0
distro-info===0.23ubuntu1
docutils==0.16
dulwich==0.19.15
duplicity==0.8.12.0
dynamic-reconfigure==1.7.3
empy==3.3.2
entrypoints==0.3
exceptiongroup==1.1.1
executing==1.2.0
Farama-Notifications==0.0.4
fasteners==0.14.1
fastimport==0.9.8
fastjsonschema==2.17.1
filelock==3.12.2
future==0.18.2
gazebo-plugins==2.9.2
gazebo-ros==2.9.2
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
google-auth==2.20.0
google-auth-oauthlib==1.0.0
gpg===1.13.1-unknown
grpcio==1.54.2
gym==0.21.0
gym-notices==0.0.8
gymnasium==0.28.1
html5lib==1.0.1
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
image-geometry==1.16.2
importlib-metadata==4.13.0
importlib-resources==5.12.0
incremental==16.10.1
interactive-markers==1.12.0
ipykernel==6.23.2
ipython==8.12.2
ipython-genutils==0.2.0
ipywidgets==8.0.6
jax-jumpy==1.0.0
jedi==0.18.2
Jinja2==3.1.2
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
jsonschema==4.17.3
jupyter==1.0.0
jupyter-client==8.2.0
jupyter-console==6.6.3
jupyter-core==5.3.1
jupyter-events==0.6.3
jupyter-server==2.6.0
jupyter-server-terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
laser-geometry==1.6.7
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lit==16.0.6
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
lz4==3.0.2+dfsg
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.1.2
matplotlib-inline==0.1.6
mdurl==0.1.2
message-filters==1.16.0
mistune==3.0.1
monotonic==1.5
mpi4py==3.0.3
mpmath==1.3.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
nest-asyncio==1.5.6
netifaces==0.10.4
networkx==3.1
nose==1.3.7
notebook==6.5.4
notebook-shim==0.2.3
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.1.0
olefile==0.46
opencv-python==4.7.0.72
overrides==7.3.1
packaging==23.1
pandas==2.0.2
pandocfilters==1.5.0
paramiko==2.6.0
parso==0.8.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pkgutil-resolve-name==1.3.10
platformdirs==3.7.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==4.23.3
psutil==5.5.1
ptyprocess==0.7.0
pure-eval==0.2.2
py-ubjson==0.14.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pyassimp==4.1.4
pycairo==1.16.2
pycparser==2.21
pycryptodomex==3.6.1
pycups==1.9.73
pydot==1.4.1
pygame==2.4.0
PyGithub==1.43.7
pyglet==1.5.0
Pygments==2.15.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
PyOpenGL==3.1.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pypng==0.0.20
PyQRCode==1.2.1
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.19.3
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-gitlab==2.0.1
python-gnupg==0.4.5
python-json-logger==2.0.7
python-qt-binding==0.4.4
python-snappy==0.5.3
PyTrie==0.2
pytz==2023.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==25.1.0
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
qtconsole==5.4.3
QtPy==2.3.1
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
resource-retriever==1.12.7
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
roman==2.0.0
rosbag==1.16.0
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosdep==0.22.2
rosdep-modules==0.22.2
rosdistro==0.9.0
rosdistro-modules==0.9.0
rosgraph==1.16.0
rosinstall==0.7.8
rosinstall-generator==0.1.23
roslaunch==1.16.0
roslib==1.15.8
roslint==0.12.0
roslz4==1.16.0
rosmake==1.15.8
rosmaster==1.16.0
rosmsg==1.16.0
rosnode==1.16.0
rosparam==1.16.0
rospkg==1.5.0
rospkg-modules==1.5.0
rospy==1.16.0
rosservice==1.16.0
rostest==1.16.0
rostopic==1.16.0
rosunit==1.15.8
roswtf==1.16.0
rqt-action==0.4.9
rqt-bag==0.5.1
rqt-bag-plugins==0.5.1
rqt-console==0.4.11
rqt-dep==0.4.12
rqt-graph==0.4.14
rqt-gui==0.5.3
rqt-gui-py==0.5.3
rqt-image-view==0.4.17
rqt-launch==0.4.9
rqt-logger-level==0.4.11
rqt-moveit==0.5.10
rqt-msg==0.4.10
rqt-nav-view==0.5.7
rqt-plot==0.4.13
rqt-pose-view==0.5.11
rqt-publisher==0.4.10
rqt-py-common==0.5.3
rqt-py-console==0.4.10
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.14
rqt-robot-steering==0.5.12
rqt-runtime-monitor==0.5.9
rqt-rviz==0.7.0
rqt-service-caller==0.4.10
rqt-shell==0.4.11
rqt-srv==0.4.9
rqt-tf-tree==0.6.3
rqt-top==0.4.10
rqt-topic==0.4.13
rqt-web==0.4.10
rsa==4.9
rviz==1.14.20
scipy==1.10.1
SecretStorage==2.3.1
Send2Trash==1.8.2
sensor-msgs==1.13.1
service-identity==18.1.0
Shimmy==0.2.1
simplejson==3.16.0
sip==4.19.21
six==1.14.0
smach==2.5.1
smach-ros==2.5.1
smclib==1.8.6
sniffio==1.3.0
soupsieve==2.4.1
stable-baselines3==1.8.0
stack-data==0.6.2
sympy==1.12
systemd-python==234
tensorboard==2.13.0
tensorboard-data-server==0.7.1
terminado==0.17.1
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.6
tf2-kdl==0.7.6
tf2-py==0.7.6
tf2-ros==0.7.6
tinycss2==1.2.1
topic-tools==1.16.0
torch==2.0.1
torchaudio==2.0.2+cpu
torchvision==0.15.2+cpu
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
Twisted==18.9.0
txaio==2.10.0
typing-extensions==4.6.3
tzdata==2023.3
u-msgpack-python==2.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urdfdom-py==0.4.6
urllib3==1.25.8
usb-creator==0.3.7
vcstools==0.1.42
wadllib==1.3.3
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.6.0
Werkzeug==2.3.6
widgetsnbextension==4.0.7
wrapt==1.11.2
wsaccel==0.6.2
wstool==0.1.18
xacro==1.14.16
xkit==0.0.0
zipp==3.15.0
zope.interface==4.7.1
